--- NuclearesWeb.Plant.SteamTurbine.cs
@@ -1,9 +1,0 @@
-﻿namespace LibNuclearesWeb;
-
-public partial class NuclearesWeb
-{
-    public partial class Plant
-    {
-        public partial class SteamTurbine { }
-    }
-}
