+++ NuclearesWeb/Plant/Reactor/Core/ControlRodBundleModel.cs
@@ -1,0 +1,16 @@
+﻿using System.Text.Json.Serialization;
+
+namespace LibNuclearesWeb.NuclearesWeb.Plant.Reactor.Core;
+
+public class ControlRodBundleModel
+{
+    [JsonIgnore]
+    private readonly NuclearesWeb? _nuclearesWeb;
+
+    public ControlRodBundleModel() { }
+
+    internal ControlRodBundleModel(NuclearesWeb nuclearesWeb)
+    {
+        _nuclearesWeb = nuclearesWeb;
+    }
+}
