+++ NuclearesWeb/Plant/SteamTurbineModel.cs
@@ -1,0 +1,3 @@
+﻿namespace LibNuclearesWeb.NuclearesWeb.Plant;
+
+public partial class SteamTurbineModel { }
